+++ admin/index.php
@@ -60,6 +60,7 @@ $expected = array(
     'ena_country', 'disa_country', 'del_country',
     'ena_state', 'disa_state', 'del_state',
     'ft_save', 'ft_del', 'ft_move',
+    'rule_del', 'rule_del_regions', 'rule_add', 'rule_save',
     // Views to display
     'history', 'orders', 'ipnlog', 'editproduct', 'editcat', 'categories',
     'pov_edit', 'other', 'products', 'gwadmin', 'gwedit',
@@ -74,6 +75,7 @@ $expected = array(
     'editregion', 'editcountry', 'editstate',
     'regions', 'countries', 'states',
     'features', 'ft_view', 'ft_edit',
+    'rules', 'rule_edit',
 );
 foreach($expected as $provided) {
     if (isset($_POST[$provided])) {
@@ -182,6 +184,12 @@ case 'savecat':
     }
     break;
 
+case 'rule_save':
+    $Rule = Shop\Rules\Zone::getInstance($_POST['rule_id']);
+    $Rule->Save($_POST);
+    COM_refresh(SHOP_ADMIN_URL . '/index.php?rules');
+    break;
+
 case 'ft_save':
     $FT = new Shop\Feature($_POST['ft_id']);
     if (!$FT->Save($_POST)) {
@@ -242,6 +250,45 @@ case 'pv_del':
         COM_refresh(SHOP_ADMIN_URL . '/index.php?editproduct&tab=variants&id=' . $_REQUEST['item_id']);
     }
     exit;
+    break;
+
+case 'rule_add':
+    $rule_id = SHOP_getVar($_POST, 'rule_id', 'integer', 0);
+    if ($rule_id > 0) {
+        switch ($actionval) {
+        case 'region':
+        case 'country':
+        case 'state':
+            Shop\Rules\Zone::getInstance($rule_id)
+                ->add($actionval, SHOP_getVar($_POST, $actionval . '_id', 'array', array()))
+                ->Save();
+            break;
+        }
+    }
+    COM_refresh(SHOP_ADMIN_URL . '/index.php?' . http_build_query($_GET));
+    break;
+
+case 'rule_del':
+    $rule_id = SHOP_getVar($_POST, 'rule_id', 'integer');
+    if (!$rule_id) {    // maybe came from $_GET
+        $rule_id = SHOP_getVar($_GET, 'rule_id', 'integer');
+    }
+    if ($rule_id) {
+        Shop\Rules\Zone::deleteRule($rule_id);
+    }
+    COM_refresh(SHOP_ADMIN_URL . '/index.php?rules');
+    break;
+
+case 'rule_del_regions':
+    $rule_id = SHOP_getVar($_POST, 'rule_id', 'integer', 0);
+    if ($rule_id > 0) {
+        Shop\Rules\Zone::getInstance($rule_id)
+            ->del('region', $_POST['region_del'])
+            ->del('country', $_POST['country_del'])
+            ->del('state', $_POST['state_del'])
+            ->Save();
+    }
+    COM_refresh(SHOP_ADMIN_URL . '/index.php?rules');
     break;
 
 case 'ft_del':
@@ -1067,10 +1114,21 @@ case 'suppliers':
     $content .= Shop\Supplier::adminList();
     break;
 
+case 'rules':
+    // Display the list of zone rules
+    $content .= Shop\Menu::adminRegions($view);
+    $content .= Shop\Rules\Zone::adminList();
+    break;
+
 case 'features':
     // Display the list of features
     $content .= Shop\Menu::adminCatalog($view);
     $content .= Shop\Feature::adminList();
+    break;
+
+case 'rule_edit':
+    $content .= Shop\Menu::adminRegions('rules');
+    $content .= Shop\Rules\Zone::getInstance($actionval)->Edit();
     break;
 
 case 'ft_edit':