+++ language/english_utf-8.php
@@ -650,10 +650,24 @@ $LANG_SHOP = array (
 'lead_time' => 'Lead Time',
 'disp_lead_time' => 'Allow %s',
 'on_sale' => 'On Sale!',
+'rules' => 'Zone Rules',
+'rule_name' => 'Rule Name',
+'sales_are' => 'Sales Are',
+'allowed' => 'Allowed',
+'denied' => 'Denied',
+'chk_all' => 'Check/Uncheck All',
+'del_selected' => 'Delete Selected',
+'del_rule' => 'Delete Rule',
+'save_rule' => 'Save Rule',
+'new_rule' => 'New Rule',
+'edit_rule' => 'Edit Rule',
 'chk_all' => 'Check All',
 'default' => 'Default',
 'cust_text_override' => 'Custom Text Override',
 'prod_features' => 'Product Features',
+'zone_rule' => 'Zone Rule',
+'restrictions' => 'Restrictions',
+'none_defined' => 'No %s have been defined.',
 'incl_nontax' => 'Include non-taxable sales?',
 );
 if (isset($_SHOP_CONF['ena_ratings']) && $_SHOP_CONF['ena_ratings']) {
@@ -685,7 +699,11 @@ $LANG_SHOP_HELP = array(
 'hlp_var_bulk_edit' => 'Editing multiple product variants at once. Any updates to the below fields will be applied to each selected variant.',
 'hlp_var_after_item' => 'Variants must be added after the item is created.',
 'hlp_bulk_cats' => 'If any categories are selected, all product categories will be deleted and replaced by the new selections.',
+'hlp_chk_all' => 'Select or de-select all items',
+'hlp_new_rule' => 'Enter a name and select the type of rule. After saving, visit Regions, Countries and States to add regions to the rule.',
 'hlp_prod_feat_form' => 'Select features to add to the product. If a custom text string is entered then it will be used, otherwise the selection from the Feature Value list is used. Check an existing feature to delete it, or add new features by clicking the (+) icon.',
+'hlp_rules_item' => 'This item cannot be shipped to the location given. Click to delete and resubmit your order.',
+'hlp_rules_noitems' => 'One or more of the items on your order can be shipped to your selected shipping address. Please edit the order or shipping information.',
 );
 
 $LANG_MYACCOUNT['pe_shop'] = 'Shopping';