+++ sql/mysql_install.php
@@ -79,6 +79,7 @@ $_SQL = array(
   `supplier_ref` varchar(64) NOT NULL DEFAULT '',
   `lead_time` varchar(64) NOT NULL DEFAULT '',
   `def_pv_id` tinyint(1) unsigned NOT NULL DEFAULT '0',
+  `zone_rule` int(11) unsigned NOT NULL DEFAULT '0',
   PRIMARY KEY (`id`),
   KEY `products_name` (`name`),
   KEY `products_price` (`price`),
@@ -672,9 +673,19 @@ $SHOP_UPGRADE['1.2.0'] = array(
       `fv_text` varchar(40) DEFAULT NULL,
       PRIMARY KEY (`prod_id`,`ft_id`)
     ) ENGINE=MyISAM",
+    "CREATE TABLE `{$_TABLES['shop.zone_rules']}` (
+      `rule_id` int(11) unsigned NOT NULL AUTO_INCREMENT,
+      `rule_name` varchar(64) NOT NULL DEFAULT '',
+      `allow` tinyint(1) unsigned NOT NULL DEFAULT '0',
+      `regions` text,
+      `countries` text,
+      `states` text,
+      PRIMARY KEY (`rule_id`)
+    ) ENGINE=MyISAM",
     "ALTER TABLE {$_TABLES['shop.products']} ADD `supplier_ref` varchar(64) NOT NULL DEFAULT '' AFTER `supplier_id`",
     "ALTER TABLE {$_TABLES['shop.products']} ADD `lead_time` varchar(64) NOT NULL DEFAULT '' AFTER `supplier_ref`",
     "ALTER TABLE {$_TABLES['shop.products']} ADD `def_pv_id` tinyint(1) unsigned NOT NULL DEFAULT '0'",
+    "ALTER TABLE {$_TABLES['shop.products']} ADD `zone_rule` int(11) unsigned NOT NULL DEFAULT '0'",
     "ALTER TABLE {$_TABLES['shop.suppliers']} ADD `lead_time` varchar(64) NOT NULL DEFAULT '' AFTER `dscp`",
     "ALTER TABLE {$_TABLES['shop.product_variants']} ADD `img_ids` text NOT NULL DEFAULT '' AFTER `supplier_ref`",
     "ALTER TABLE {$_TABLES['shop.product_variants']} ADD `dscp` text NOT NULL DEFAULT '' AFTER `img_ids`",
@@ -700,5 +711,6 @@ $_SQL['shop.countries'] = $SHOP_UPGRADE['1.1.0'][7];
 $_SQL['shop.features'] = $SHOP_UPGRADE['1.2.0'][0];
 $_SQL['shop.features_values'] = $SHOP_UPGRADE['1.2.0'][1];
 $_SQL['shop.prodXfeat'] = $SHOP_UPGRADE['1.2.0'][2];
+$_SQL['shop.zone_rules'] = $SHOP_UPGRADE['1.2.0'][3];
 
 ?>