+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Release TBD
   * A subset of product images can be assigned to each variant.
   * New reorder report listing items at or below their reorder qty.
   * Allow plugins to use the catalog display by calling index.php with `category=pi_name`.
+  * Restrict product sales by region.
 
 ## v1.1.2
 Release 2020-02-15